+++ bin/setupOperation/database.py
@@ -467,3 +467,31 @@ class initialize(object):
 				description='The HDFS blocksize in bytes. Can usually be found in /etc/hadoop/conf/hdfs-site.xml (search for dfs.blocksize)')
 			self.configDB.execute(query)
 
+		if result_df.empty or (result_df[0] == 'sqoop_import_default_mappers').any() == False:
+			query = sa.insert(configSchema.configuration).values(
+				configKey='sqoop_import_default_mappers', 
+				valueInt='12', 
+				description='How many mappers should be used for tables who have never been imported before')
+			self.configDB.execute(query)
+
+		if result_df.empty or (result_df[0] == 'sqoop_import_max_mappers').any() == False:
+			query = sa.insert(configSchema.configuration).values(
+				configKey='sqoop_import_max_mappers', 
+				valueInt='32', 
+				description='The maximum number of mappers to use during imports')
+			self.configDB.execute(query)
+
+		if result_df.empty or (result_df[0] == 'sqoop_export_default_mappers').any() == False:
+			query = sa.insert(configSchema.configuration).values(
+				configKey='sqoop_export_default_mappers', 
+				valueInt='2', 
+				description='How many mappers should be used for tables who have never been exported before')
+			self.configDB.execute(query)
+
+		if result_df.empty or (result_df[0] == 'sqoop_export_max_mappers').any() == False:
+			query = sa.insert(configSchema.configuration).values(
+				configKey='sqoop_export_max_mappers', 
+				valueInt='32', 
+				description='The maximum number of mappers to use during exports')
+			self.configDB.execute(query)
+